requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
Pillow==10.2.0

